Fixed two more bugs
month is 0-indexed for some reason so April is displaying as 3. added 1
when displaying in task chart.
if user typed in date instead of selecting from calendar, it is not
updated unless user presses enter while in textbox, which doesn't make
sense. Added listener that updates the datePicker date when user exits
focus from the date textbox